Starting on the account service
This service doesn't seem all that useful as it won't be good to put into a composite with the shopping cart service :(.  I think I'll follow the example of Callista Enterprise blog series by creating the product, recommendation and review services along with a product-composite service.  Seems like a good starting point to provide some good examples and unit tests and yada yada.